cctv 레이어 추가 / 메인페이지 재구성
cctv 레이어 추가 / 메인페이지 재구성 (js, css -> community페이지 기능, myplace페이지 기능), community페이지 삭제, myplace 삭제, 서버단 코드 정리

+++ src/main/java/mapdiary/map/service/MapService.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface MapService {
     LocationVO insertLocation(LocationVO vo) throws Exception;
     List<LocationVO> selectLocationList(LocationVO vo) throws Exception;
-    void updateSharedStatus(int locationId, boolean isShared);
+    void updateSharedStatus(Long locationId, boolean isShared);
 }
